moved AutoProps
