+++ packages/theatre/src/lib/sheetObject/autoProps/AutoProps.svelte.d.ts
