+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/defaults/color.ts
