+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/defaults/degrees.ts
