+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/defaults/euler.ts
