+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/defaults/generic.ts
