+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/defaults/normalized.ts
