+++ packages/theatre/src/lib/sheetObject/autoProps/transfomers/getDefaultTransformer.ts
