+++ packages/theatre/src/lib/sheetObject/autoProps/utils/getInitialValue.ts
