+++ packages/theatre/src/lib/sheetObject/autoProps/utils/isPrimitive.ts
