+++ packages/theatre/src/lib/sheetObject/autoProps/utils/makeAlphanumeric.ts
