+++ packages/theatre/src/lib/sheetObject/autoProps/utils/parsePropLabel.ts
